Some misc code Clean up. Combine the Check for full spoiler log with Check for spoiler log. Check for full with an argument. Made the prompt save able to ignore whether or not there are unsaved changes. Currently not used. Probably some other stuff.
Added a check to the function that unchecks item that will stop it's randomized item from being set to unaquired if that item is also the randomized item of another check. I thought this would be useful when I made it, it's probably not but I left it and put it as a StrictLogicHandeling Function.

Made the Item select list Sorted, Not sure why it wasn't.

Added a save prompt when closing the form.

Massive Update to the OOT Dictionary, Basically Have to do it by hand when I have time but I'm sure eventually It'll get done.