Me me BIG BOY Update
Added a fully functional Logic Editor. This was originally made to help with making the OOT logic since I can't edit that in the MMR randomizer, but it turned into what I beleive to be an improved version of what is offered in the MM Randomizer.

Added Time Available/Needed Data to the logic object. Mostly just for compatibilty with the logic editor but also so I can eventually make the tracker sort checks by time of day.

Allowed the recreate logic function to use a settings file.

Made the Logic objects class public. Easy fix for someting that caused so much headache.

The dictionary finder now needs to be passed a version instead of allways using the main version used by the tracker aming it a bit more versitile.

Updated a TON of item names on the OOT Dictionary.

Huge change to the item select form. It can now use a checked item list for certain functions which IMO is cleaner that ctrl selecting things. The code is also just generally nicer.

The item select list now remembers what was selected if you change the filter. I can't believe this never came up as an issue before.