Save the logic in any way should now update that there are no unsaved changes.
Fixed the logic editor generating white space at the end of the file casuing it to error when loaded into the Radnomizer.

Added the Icon to all Forms.

Added the option for the logic editor to show dictionary names instead of The tracker names.

Added the option for the logic editor to show Spoiler log names instead of tracker names.

Added Save Functionality to the logic editor.

Various code cleanup.