Added a local config file that can be used to store information after the application has closed.
Added a first time setup process.
This process will only run if an options file has not been created. It will show the user the about page and have them configure some options for the options text file.

Added some functions to wreap text in a list box by checking if a string will fit in the list box horizontally and if not splitting it into multiple strings that will. Implimented this in the about page.

Cleaned up the about page code by implimenting the above feature. Also added information on the newly added Features.

Added a function that will create a divider that matches the width of the list box. Implimented it in the main interface and about page.