Entrances will now be grouped underneath items in list boxes that contain both.
Marked items/entrances will now be moved to the bottom of the list.

Removed the selected items list box from the randomization options page. It was pretty unstable for some unknown reason.

Lots of internal code clean up. Renaming of object and adding extetions to the logic entry object for commonly used checks.