Fixed the pre entrance rando area clear function crashing the tracker. This is what happens when you don't properly test massive Code Changes.
Fixed the playthrough generator not showing the correct names for boss kills.

Made unrandomized items be treated like fake items in the playthrough generator and What unlocked this function.

Added an option for moving marked items to the bottom of the list.

Changed the entrance rando version to 16. it was previously 14 but a new pre entrance rando version 14 has been release. I imagine by the time entrance rando comes out its logic version will be far beyond 16.

Added support for logic version 14 which adds Ocarina of time and song of time to the item pool.

The item select list for locations and entrances will not show items that are not randomized