Added the ability to load saves made with tracker versions before the big rewrite.
Fixed the update names function.

Made some changes to the pathfinder that will hopefully resolve a few edge cases where a path could not be found to certain entrances if they use fake items that require other entrances. This is still a problem, this fix might resolve 1 out of every 10 of these issues.