The path finder now runs async which will prevent it from locking up the tracker when doing calculations. Pretty pointless now that it only takes  a second or 2 at most.
Added some extreamly awful code to fix a dumb problem with the path finder. basically now every "Find path" operation stores it's results seperatly in a list and listboxes are told which set of paths to look at in a list. This allows two or more pathfinders to be open at the same time and not fight over the global paths variable. It also removes any danger of two pathfinders running at the same time now that they run async. The downside is the tracker now stores every path that has been generated since the program opened. Whatever, it works, don't @ me.