The pathfinder will now skip the pathfinding process if the destination is in the same area as the exit.
Fixed the printing of paths to properly display when song of time is being used.

Exits from majoras lair will now be ignored. There are no actual exits from majora but logic uses it to mark inaccessabel areas.

Updated the song list to include Sun Song as well as the two song of time variations.