Moved all the heavy lifting of creating and writing filters to dedicated functions.
Added functionality to the remaining buttons.

Whether or not entrances are being displayed in the item box will now be taken into account when writing filters. If entrances are written in the item box but are not selected in the map screen, the filter will be set to ignore entrances. The same is done with items.