The context menu is now recreated when the "format form" function is run allowing for more dynamic options that can change along with other options.
The entrances and Locations list box now have seperate context menus and slightly different options.

The contex menu will now no longer appear when rightclicking an item that can't be interacted with.

Fixed a crash when right clicking a list box over empty space with no selected items

Added a "Navigate to" option to the context menu that will fill the destination of the pathfinder with the selected object. This option will only appear in the entrances list box unless "Show items as destinations" is true.