Added a function to return the lacation the given item was palced at.
Added support for Entrance Pairs files for other Game. The naming convention is the same as dictionaries (GameCode + "ENTRANCEPARISV" + VersionNumber)

Added som functions to the Item Display. It can now:
Click on an item to filter for it in the main tracker.
Display Counts of items as well as if it's marked and not checked.
Invert the colors of an image (Used to create the double defense Icon)
Item display is still far from done.

Orginized the recources folder into sub folders as it was getting a bit crowded.