The "What unlocked this" fucntion now displays what fake items were used to unlock the check to give a better idea of what is expected.
The default button in the change font window will now update the values in the Numberic updown and combo box.

Made listbox item text change to white when it's highlighted.

made the listbox item height a bit bigger to prevent certain font sizes from cutting off the bottom of letters.