The tracker will now attempt to generate a playthrough even if spoiler data is missing.
The "What unlocked this" function is now far more thourough in the data it shows. It will show:
What logic entries were used to unlock it
What real items were used to unlock it
What fake items were used to unlock it as well as what real items were used to unlock those fake items.

It is also now displayed in a proper list box rather than a message box.