Started added support the trick list recently introduced in dev versions of the MM rando. The tracker can now recognize entries that are marked as "tricks" when importing a logic file. if a trick is not enabled, any conditional containing that trick will be ignored when calculating logic. All tricks are enabled by default when creating logic.
Currently there is no way to edit which tricks are enabled. This functionality is planned for the near future.

Previous version of logic that do not contain trick data will still work as usual.

Fixed the item display filter function to use the correct symbol now that _ has replaced the functionality of &.