Added trick functionality to the logic editor. The logic editor can now mark fake items as tricks and set their trick tooltip.
Added the ablity to delete unused fake items from logic. If the item is used in any entries it will display those entries.

Added a "What is this used in?" tool to the logic editor that will display all entries that use the selected item.

Added an option to the logic editor to save logic without the extra line for trick information. Logic that has trick data will not work in versions of the randomizer that do not have tricks. This is allows for backwards compatibility.

The go back button will now work if a value is manually entered into the numberic up down.