Made the active port remove it's self when the server stops listening.
Added the ability to save and load your Sending IP list. This list will also save your current IP so others can load your save to mimic your IP list. If the IP address belonging to the person loading the file is in the list, or if the save contains an address already present in the ip list, it will be ignored.

Added the option to force net data to only be allowed to mark items. If a packet contains items to be checked they will be marked instead.

Added a toggle for auto port forwarding for cases it is not desirable.

Added the option to automatically add IP adresses you recieve data from to your Sending list if they don't already exits.

Added a dev option to paste the net data that would be sent to your clipboard.

Added the source IP address and port number to the packet that is sent to other players. This allows that data to be auto added to their instance if the option is enabled.