What unlocked this now displays the check is not available if the check is not available.
Fixed the logic editor nud index adding to the go back list if the value tries to change but doesn't change.

Made the fake item options in the logic editor ignore fake items added by the randomizer

Added an option to the check locations list box to uncheck a location and automatically set the item after unchecking it.

Fixed the pathfinder crashing when trying to print a path due to the fact that the draw function now considers paths and item entries the same and tries to run the same checks on pathfinder items that it does logic items.

Added the ability to set the latest destination as the starting location in the path finder by double clicking the "Destination" label.