Added some simple code to handle Ice Traps in the spoiler log.
If they are detected when reading a spoiler log it just marks the item in that location as "junk" (ID -1). Hopefully that won't introduce any strange behavior since the "Junk" item has been around since the start of the tracker, just hasn't been used much.

Made the function that checks for spoiler data count the "Junk" ID (-1) as a valid spoiler entry. I don't think this will be an issue, but if it is I added a toggle in the function to ingore them if need be.