Added Code to support MultiWorld.
This is untested and might not work right, but the method should be solid. I also have no idea what multiworld will look like in MMR but as long as the spoiler log detailes which item belongs to which player I should be able to make it work with this method.

This will not function at all at the moment as certain settings can not be enabled. GUI elements to do the following still need to be added.
-Enable MultiWorld
-Select your Player ID
-Select which Player an Item should go to if a spoiler log is not applied