Added a very basic encryption function.
For now it's unused. This may be applied to the net data that is sent between clients in the future. This encryption is very easy to break but if a person knows how to do that, they can probably also find the IP data listed in the packet from other sources.

I don't usually copy code striaght off of github but I couldn't be bothered for this. Source:
https://stackoverflow.com/questions/202011/encrypt-and-decrypt-a-string-in-c