Starting Items will now always appear in the checked locations Box.
If an item is obtained but has not bee assigned to a locationt, it will appear in the checked item box as being obtained from "Unknown". It can be double clicked to be set as unobtained. This should never happen, but now if it does it can be easily fixed.

"Multiworld" and "Starting" have been added to the categories.