Work on the OOTR support.
The dictionary is so close to done at this point, a large number of MQ checks are misnamed and there aren't always enough of certain items, but it is 100% playable (aside from MQ)

The Spoiler log converter is now much smarter about adding checks missing from the OOTR spoiler log. Almost everything should be covered even with minimal settings.

Unfortunetely due to the massive changes to the Dictionary, the included logic file is completely broken and will need to be done from scratch.