Finished Wind Waker Rando Logic And Dictionary.
WWR support should now be done aside from some possible logic bug fixes in the future. The first live test will happen soon.

Ensured the applications current directory is always the same as the base directory in case the application is opened from a MMRTSave file elsewhere on the system.