Finished the logic editors "create (any X of A,B,C,D)" Function. This can create conditionals that represent something like "Any 3 of the 7 forest temple keys". These entries can be very large so a cap was put in place to stop the editor from attempting to create entries that would crash it.
Added a "Logic Parser" that can take "Code" formatted logic such as (A&(B|(C&D))) and convert it to a set of conditionals.

The Logic Editor will now refuse to display conditionals that would be large enough to crash the editor.

OOTR Dictionary: removed the extra Key entries as they were not neccissary.

OOTR Logic: Added Key Logic