"SpoilerLocation" and "SpoilerItem" are now lists of possible names the spoiler log could use.
The names are seperated in the dictionary by |

Added to prepare for progressive items in MMR. This may also allow for the use of only one dictionary in the future. It will defenitely help with the fact that OOTR changes spoiler names just about every update.

This is untested and will most likely break something.