Added a setting to enable progressive items in logic.
With this setting enabled, whenever logic is checking to see if a progressive item is available, it will instead check to make sure you have enough of that progressive item to reach the correct stage. In other words, user logic does not need to be changed.

Item names will also automatically be displayed as "Progressive X" instead of their original name.

Importing spoiler logs will now automatically apply the setting found in the spoiler log to the tracker.