Added a method in logic to create an "any X of A,B,C,D" Entry. It works as follows:
If an entry whose name starts with MMRTCombinations is found in the requirements for an item, it will instead use this method to determine whether The item is avalable. Example: "MMRTCombinations16"

It will take  the number imediately following "MMRTCombinations" and set that as the "Needed amount". It will then check if the number of aquired conditional sets in the items conditionals is greater than or equal to the "Needed amount". If so the item is marked as available.

This can be used to add a "Any X of A,B,C,D" where the amount of conditionals would be to large for the tracker to handle.

For example, Any 50 of 100 skulltullas would need 100,891,344,545,564,193,334,812,497,256 conditionals to be represented using the default logic, whereas it can be added using this method with 1 requirement and 100 conditionals.

Added a use case for this new system to the skulltula reward logic in the OOTR dictionary.