Made the playthrough generator and seed checker a bit more stable.
The playthrough generator and seed checker will never need to "Uncheck" and item, so the code to handle that was removed. This also fixed some strange infinite loops with OOTR Logic.

The playthrough generator will now display "Beat the game" instead of "Defeat majora" when the istance in not MMR.

Added an option to make the seed checker tell you where each item was located in adition to whether or not it's obtainable. This is currently hardcoded to be OFF.

Re enabled Seed Checker and Playthrough generator when other games are used. Hopefully the weirdness was worked out.

Fixed some misc things in the OOTR spoiler log converter and OOTR files