Added a dedicated function for checking an items availablity. Like the "IsUseable" function this helped clean up the code quite a bit.
Dungeon area clear now works in a more sensable way, closer to what the actual randomizer does.

Added a method for logic to check if a specific item is at a specific check.
If an entries Requirements contains a fake item named "MMRTCheckContains". it will return true if the item check, identified by the first requirment is the list (besides the indicator), contains any of the items in the first conditional. It will then check the next requirement and conditional and so on until it finds a set that returns true or reaches the end and returns false.

Moved all Logic Object extentions out of LogicObjects and into their own dedicated class file.