The Logic parser is now a multiline textbox.
The logic parser will now ingore comment lines that start with #. The comment must be at the end of a line or on it's own line.

The logic parsers cursor will now keep its postion when ading an item via the list veiw.

Renamed a lot of areas in the WWR dictionary to better represent the actual location.

Small fixes to the OOTR logic