Changed the logic parsers method of converting names to IDs. It is now much more reliable.
The logic parser will now recognise when an items dictionary name is listed in the logic box and add that item to the used items list.

Holding ctrl when clicking the convert names button in the logic parser will now convert ID's into dictionary names.

Unfortunetaly with these changes, logical operators other than & | ( ) are a bit finiky, the parser will attempt to convert them to the standard operator when the "Convert names" button is used.

When populating a tracker instance, if that instance already has dictionary data, it will not attempt to find a dictionary in the dictionary folder.

Began preparing files for support of the skyward sword rando, This one looks like it will be super easy, like everything including logic can be created automatically easy.