Finished files for Skyward Sword Rando support
Everything is done, logic included. All created automatically trhough code.

This rando is still in it's infancy and I'm sure it will go through a lot of changes, but hopefully I was able to make a system that can import everything straight from the source files when it does change.