Some overhauls to the Skyward Sword Rando Files
Removed a lot of error checking in logic creation, if something errors that means something needs fixing and I don't want to just skip over that.

Added the MMRTGameClear entry to the SS Logic.

Options are now read correctly from the spoiler log

Added an option for closed thunder head

Added entries for the seeds required dungeons. THis data is extracted from the spoiler log and taken into consideration when generating a playthrough.

Unimplimented checkes now don't hacve a location entry, only item. We still want the item in the pool but don't want the location to show on the tracker.

The randomizer is updating as fast as I can add support for it. There were already some changes since the last commit. It will be awhile before this is finalized, but at least I have a foundation and can do some testing.