Overhauled Item strings in the Randomization options code.
Improved overall stability and ensured the item string logic matches the Majoras Mask Randomizer Exactly.

When a setting chnages, the custom Item string will now be generated and pasted into their designated text boxes based on your settings in the randomizer.

Added support for the starting item string. This was a little hard to impliment and will have to be updated manually if new starting items are added unfortunately.

Starting items will be imported along with other setting from an MMRT setting.json or spoiler log.