Changed teh Seed Checker form into a general spoiler log tools form that contains Seed Checker Playthrough Generator Spoiler Log Lookup
Added a new feature that will show you what sphere an item can be obtained in. This will give you an idea of how far into the seed the item is with telling you exactly where it is.

You can now manaully select what item is needed for game completion in the playthrough generator.