The Playthrough Generator will now mark items you have obtained in your current playthrough.
Display Total counts for Enabled and disabled tricks as well as an overall count for all tricks and logic items.

Added support for the new category system in 1.14. Settings files will now work if you are using the checkboxes instead of the custom item list.

Added Gossip Stones as Item checks. They will only be something marked, you should never have to assign an item to them.

Settings import will handle gossip stone entries based on whether they exist in logic or not they are set to have hints in your settings.

Handle a strange bug with the tool that ensures the logic editors create permutation function doesn't oveload the program with to many entries. It's still not working 100% correctly but it does it's job.