Fixed some of the massive lag introduced in newer versions.
The lag only happens with newer version of the MMR logic and most likely has to do with the massive additions to logic in the newer version of the randomizer (1000+ checks)

This is done via a change to the check availability system that makes it only calculates the availability of checks if they contain an item that was aquired/unaquired in the action that triggered the check availability to happen.

The lag still exists most times logic is calculated such as when creating it for the first time or changing setting such as progressive items.

It is however almost completely gone when checking an item on the main interface which is 90% of what you do in the tracker.

I'll eventually hopefully expand this to other functions in hopes of speeding things up further.

Needless to say this is very untested and due to it's nature extreamly untested. Due to where this code is executed it won't touch any special logic calculations such as Random prices wallet swapping and dungeon entrance swapping so it should hopefully be pretty stable.