Updated Item Display code to use the new Logid IDs. Another thing broken by that update :)
Fixed a bug with the Logic calculation speedup method that tries to skip items that don't need updating. It was not taking into account unrandomized items as something that could cause an entries availablility to need updating.

Added a check to the same speed up to always update an item if it has a prgressive item in it's requirements or conditionals. I'm not 100% sure this is neccesary but it's a good saftey net and not a huge performance hit.