Price data will now apply to checks that can have multiple prices.
If any check has one or more requirement or conditional that are a fake item with a price, those price are all printed when the loaction is marked.

The spoiler log parser will now convert entries detailing multiple prices fo the same location and add them to the approprate fake item which should be in every logic file.

Removed the unused encryption code that was supposed to be for net data.

Added a bunch of copied code from the Randomizer that will allow me to copy the Randomizers item list into this code without any changes. I can then hopefully use the randomizers item list to check for any flaws in logic dictionaries.