Added support for Gossip Stone hint Text
When Checking a gossip stone, you will be asked to input the hint it gives. This hint will be shown allong with the gossip stone in the checked item box. The hint text will be wiped if the Gossip stone is unchecked.

Hints will be automatically imported along with a spoiler log and will persist even when the hint is unchecked.

Hint text can be made to persist by adding $ to the front of the text. The "remove spoiler data" option in the right click menu can be used to remove hit text.

Gossip stones can not be "Set". Setting the item will perform a full check.

Since the read Spoiler log function returns an object now, settings data is returned by the funtion to be applied to logic in the applySpoilerLogToLogic function. This doesn't change any functionality but makes things a bit cleaner.

The tracker will now prompt you before removing all spoiler data when the option is selected to prevent accidental Removal.

Price data an Gossip Stone hint data are now removed when the remove spoiler data option from the right click menu or the Remove all Spoiler data option in the Tools menu are used.