Change how randomized Price data was assigned from the spoiler log.
The SpoilerLogPriceLogicMap text file can now be used to completely customize what logic items are assigned each random price value from the spoiler log. By default it's set up to handle it exactly like the randomizer does, but since the randomizer hadles it in a kind of conveluded way, this allws you to customize how the tracker hadles it.

TheConfiguration is very robust allowing for things such as getting the lowest price of multiple entries and ingoring entries based on certain items being in logic.

Allowed fake items to be added to the DicNameToID dictionary, hopefully this doesn't bite me later.

Fixed a bug where dungeon entrance clear checks would never become aquired if dungeons were not randomized.

Fixed

added method for easily removing and adding requirements and coditionals on the fly.

The tracker will now remove the letter to kafei and pendent of memories from the logic for Old Lady and Big Bomb Bag purchase. this is only neccesary during randomization and these items are not actually required to do those checks. TODO make this toggleable

Clicking the lables for the item/location/dictionary name in the logic editor will copy that name to the clipboard.

If the tracker find a duplicate values in the spoiler log price list, it will automatically add a number after the name instead of just skipping it

When a different logic is applied to the instance, Item locations, Gossip hints and Price data are now reapplied from the spoiler log instead of being transfered from the old logic. this is because price data specifically can be assigned to different checks based on what items exist in logic.

The spoiler log is now saved in the instance when imported, right now the only use this has is reapplying the spoiler log when logic changes, but other code may benifit from it in the future.