Made saving logic as JSON default in the logic editor.
Fixed the WriteLogicToJson crashing if requirements or conditionals were null.

Fixed Json format options so the tracker outputs a json file in the exact format as the randomizer

Added the TimeSetup value to logic. The randomizer uses this for something, not sure what yet.

Added check boxes for TimeSetup to the Logic editor form.

Logic options now display location name follwed by item name to mirror how it is in the randomizer. Displaying the dictionary name as well is a togglable an option

The save file is now a little more readable. Not that you would ever need to read it. It's also 150,000+ line long so...

Made the PrintLogicObject display logic in Json format.