Cleaned up most of the code for the playtrhough generator.
Fixed a wierd bug with Dungeon Entrances and clears in playthroughs.

Starting items are now display as obtained in sphere zero and the check they are assigned to in game is never considered necessary.

Changed the are clear entry in the playthrough generator to list "Kill Boss -> Clear area" instead of "Kill boss -> Boss remains". This makes a bit more sense since with the addition of boss remains as it will be a sperate entry and are clear will now only show up if clearing the area is required for an item.

Fixed a crash when applying the logic from the logic editor to the tracker if the logic editor had JSON logic. It would try to apply the JSON logic to the tracker as non JSON logic which would mess all kinds of things up.