Fully reimplimented undo and redo.
Fixed the redo function to use the new methods.

Moved the undo and redo lists out of the tracker instance. A new object now exists to hold undo redo data. Two instances of this object now exist as global variables, one for the main form and one for the logic editor.

Gave the logicEntry object default values. Should hopefully help prevent crashes in odd cases.

Added a function to set requirements and conditionals that have no values to null. This helps prevent some bugs with other funtions that expect requirements and conditionals to be null if they have no data. This bug completely broke CleanLogicEntry so currently this function runs at the beggining of that function.

When writing an entry in the Requirement Check form, it will only attempt to clean logic if the logic was edited.