Misc Code Cleanup
The nullEmptyLogicItems function now runs during logic creation and everytime logic is calculated.

If Keysy is enabled, PerformLogicEdits will attempt to remove any keys from any location that have had their logic editied along with checks that were found to contain keys in logic. This assures it doesn't miss any checks that changed their "has keys" status during logic edits. This removes the need to always add dungeon clear items in the GetChecksNeedingKeys function.

Changed logic entries with conditionals that contain only disabled tricks to always return false.
The randomizer does the opposite, returning true if the conditional only contained tricks and all those tricks were removed. This works for the randomizer because it assumes the conditional will always have at least one entry with no tricks to fall back on.  In the randomizer, if a check is unobtainable it is given moon access as a requirement, or conditinal if tricks can make it obtainable.  It has to do this since, while moon access does make sure the location can't contain something needed to beat the game, it is eventually obtainable. The randomizer errors if any check is unobtainable, even if the check is not needed to beat the game. The tracker doesn't care. By having the tracker return false if all conditionals are inavlid tricks it gives the tracker an option to make checks unobtainable.  Since the randomizer expects logic will always have at least one conditional with no tricks to fall back on,  this will only cause inconsistancy if logic was used that is invalid in the randomizer anyway.

Removed unused code "ProgressiveItemsAquired". Not sure why it existed.

When a locations availablility is checked, it conditionals are first sorted from smallest to largest. This will hopefully prioritize easier options for obtaining the item when calculating a playthrough

Fixed bug in the MarkAreaClearAsEntry code in the playtrhough generator where the randomized snowhead clear location name was being assigned to great bay and ikana clear.

Fixed the Boss Keysy menu toggle toggling the option based on the state of small Keysy