The Logic Dictionary is now a JSON object and contains far more data than before.
The Logic dictionary is now contained within an object that hold new information such as the Default wallet capacity and Logic format it was made for. It still contains the Logic version game code and all dictionary data.

The SpoilerLogPriceLogicMap and WalletValues Text files have been deprecated as the data in the files is now contained within the logic dictionary.

Small and Boss keys are now defined in the dictionary rather than in code.

The Entrance Area dictionary is now created from data withing the Logic Dictionary. If a dictionary entry is given is given a value in the "GameClearDungeonEntrance" entry, the tracker will add the entry and data to the Entrance Area dictionary

Added the gossip stone names that the randomizer defines for each item and location. Hopefully in the future this will allow for auto marking items when gossip stones are checked.

The type of logic the tracker is reading is no longer defined by a bool of "Json or not Json". it is now defined by a string representing the file type.

Added the ability to distinguish a fake item as a "Randomizer defined fake item". This can be used to prevent the logic editor from moving, renaming or deleting the item

Fixed GreatFairyDoubleDefence not being mapped to the correct item in the item display