Progressive items are now defined Via the Logic Dictionary.
Progressive item data is now saved in the item entry it's self allowing for easier parsing of progressive items.

The code to ignore availablility updates of checks that don't require items that were updated will no longer run if the check contains any progressive items. This prevents some inconsistencies.

Removed Item parsing code from gossip hints, the hints are to vague to be reliable.

Since Tracker created fake items now exist in the Logic Dictionary, that data is now used to determine if an entries name and poistion can be edited in the logic editor.

Re added progressive item names to the Spoiler item array in the dictionary.