Rewrote pathfinder form to handle multiple pathfinder instances a lot cleaner.
Pathfinder path lists are now contained to each pathfinder instance instead of being saves statically. A new static pathfinder instance exists for the main form to handle it's pathfinder.

The option to toggle using Song of Time in the pathfinder has been removed. Instead, if the pathfinder is unable to find a valid path, it will check again using Song of time as a valid option.

If an entrance pair is checked, it is now added to the list of LastUpdated checks which prevents a bug with the performance code.

txt logic now handles fake items correctly from the JSON dictionary

added an actual button for swapping start and end paths in the path finder.

Fixed starting item string parsing with the new JSON logic dictionary system

Added Logic Dictionary for Entrance Radno 1.13 update

Marked valid Starting items in the Entrance Radno V1.12 Logic File