Added support for the Skyward Sword randomizer.
Included the dictionary file along with both Casual and Glitched logic.

Spoiler logs are currently not supported.

Added a method for the new JSON logic to carry game code information. It will be left out of MMR logic files by default so it shouldn't effect Logic edited by the Tracker being used by the randomizer. The randomizer should ignore it anyway since it's a JSON import.