Added Spoiler Log Support for Skyward Sword Rando.
The spoiler log and option parsing are now handled completely within the SSR Class file. When importing a text spoiler log, if the game is not MMR it will attempt to parse the file using the Glabal handler Class only.

Altered some code to handle an item being a starting item, but not being randomized to any checks in the game, Something that MMR never runs into.

Starting Items will now be displayed even if they are items with no original location.

"Option" checks are no longer displayed in the playthrough.