If an item can be purchased from multiple locations, the tracker will now only display the price of that item from places you are logically able to purchase it.
For example, if you check the Clocktown tingle map purchase from North Clock Town but have no way of accessing Ikana to see the price for the map there, only the clock town price will be displayed when the location is checked or marked.

This will help prevent mild spoilers incase you don't want to know the price of locations you can't check.