Checking a gossip stone now markes the location and item in the hint.
The location will only be marked if its spoiler log data says it contains the item listed in the hint. This ensures that the incrrect location will not be marked if the hinted item or location is vauge.

Made the price hiding on checks a bit stricter. Now, if a check is not available it will never return a price under any circumstances.