Add a special item check type that, when checked, can be set to a user defined number instead of an item.
-This item can be initialized by setting it's
ItemSubType = "MMRTCountCheck"
-When checked it will ask for a numberic input which it will display in the checked list.
-For now the only use for this is in conjunction with the new MMRTCombinations functionality "MMRTCombinationsDynamic"

-OOTR Logic
Starting to piece things together. A list of all locations and entrances with their vanilla items now exists and logic is able to be parsed from the web into a readable format. Now I just have to sort through all the edge cases

To Use "MMRTCombinationsDynamic"
-Create a fake item with the dictionary name "MMRTCombinationsDynamic"
-Place this item in the requirements of a check
-Place an "MMRTCountCheck" entry in the requirements of the check
-Fill the conditionals like you would with a normal "MMRTCombinations" check

The tracker will use the number data from the "MMRTCombinations" check instead of the number at the end of a typical "MMRTCombinations" to determine how many conditionals are needed