Added full support for wind waker rando.
A dictionary and logic file have been added to the appropriate folders and a function for parsing the spoiler log has been created.

This supports all current features, but will not support enemy rando if that is ever fully implimented as the logic already in place for it is to complex and can not be handled elegantly by the tracker.

Premade logic for other games will now appear in the preset menu if other games are set to allowed in your option file (so if you have opened a non MM game at least once)