The requirement checker will no longer show tricks as obtained if they are disabled.
Increased window size of the requirement checker and made the horizontal scroll bar work.

Ocarina of time rando will now import hint data from the spoiler log.

Changed the OOTR logic entry "logic_rules != glitchless" from a logic trick to an option. Specifically "logic_rules" is now an option with the choises "none, glitched and glitchless" this is how it was intended to be but since "logic_rules != glitchless" starts with "logic_" it was parsed as a trick.

When importing OOTR spoiler logs, "shuffle_dungeon_entrances" and "shuffle_overworld_entrances" are now given a default value since they don't appear in the spoiler log if logic is glitched.